refactor: colocate MoveNode and MoveNodeProxy
